Partial revert of "vagrantfile"
This partially reverts commit 46aba1eb16834415093ba18c2fa1c3c3b6b9398f.

This commit changes the installation to, instead of using the public key
included in the repository, to fetch the key from a keyserver by key ID.

Fetching a key by key ID is not secure: key IDs, despite the name, are not
unique. (The ID is 32-bits; it isn't long enough to be unique.) Moreover,
someone appears to have uploaded a key with a duplicate key ID to this
particular server, for this particular key ID, named
"Totally Legit Signing Key <mallory@example.org>". Let's not use that key.